chore(codegen): refactor end
