+++ packages/codegen/src/core/modules/JavaScript/JavaScriptCompiler/class/index.ts
@@ -1,0 +1,5 @@
+import { classSerializer } from './serializer';
+
+export const classModule = {
+	serializer: classSerializer,
+};
