+++ packages/codegen/src/core/modules/JavaScript/JavaScriptCompiler/context/index.ts
@@ -1,0 +1,9 @@
+import { getInitialContextShape } from './core';
+import { contextSerializer } from './serializer';
+
+export const context = {
+	contextSerializer,
+	functions: {
+		getInitialContextShape,
+	},
+};
