+++ packages/codegen/src/core/modules/JavaScript/JavaScriptCompiler/dictionaries/core.ts
@@ -1,0 +1,49 @@
+import { BasicActionDictionary, BasicStateDictionary } from '@yantrix/automata';
+import { TExpressionRecord, TStateDiagramMatrixIncludeNotes } from '../../../../../types/common';
+import { context } from '../context';
+import { expressions } from '../expressions';
+import { imports, TDependencyGraph } from '../imports';
+import { TDictionaries } from './types';
+
+export function getObjectKeysMap(dict: Record<any, any>) {
+	const obj: Record<string, string> = {};
+	Object.keys(dict).forEach((key: string) => {
+		obj[key] = key;
+	});
+	return obj;
+}
+
+export function setupDictionaries(props: {
+	dependencyGraph: TDependencyGraph;
+	diagram: TStateDiagramMatrixIncludeNotes;
+	stateDictionary: BasicStateDictionary;
+	actionDictionary: BasicActionDictionary;
+	expressions: TExpressionRecord;
+}) {
+	let dictionaries: TDictionaries = [];
+	dictionaries.push(
+		`export const statesDictionary = ${JSON.stringify(props.stateDictionary.getDictionary(), null, 2)}`,
+	);
+	dictionaries.push(
+		`export const actionsDictionary = ${JSON.stringify(props.actionDictionary.getDictionary(), null, 2)}`,
+	);
+	dictionaries.push(`const reducer = {${context.contextSerializer.getStateToContext({
+		diagram: props.diagram,
+		stateDictionary: props.stateDictionary,
+		actionDictionary: props.actionDictionary,
+		expressions: props.expressions,
+	}).join(',\n\t')}}`);
+	dictionaries.push(`export const functionDictionary = new FunctionDictionary();`);
+	dictionaries.push(`functionDictionary.register(builtInFunctions);`);
+	dictionaries.push();
+	imports.functions.checkForCyclicDependencies({
+		dependencyGraph: props.dependencyGraph,
+	});
+	dictionaries = expressions.functions.registerCustomFunctions({
+		diagram: props.diagram,
+		expressions: props.expressions,
+		dependencyGraph: props.dependencyGraph,
+		dictionaries,
+	});
+	return dictionaries;
+}
