+++ packages/codegen/src/core/modules/JavaScript/JavaScriptCompiler/dictionaries/index.ts
@@ -1,0 +1,11 @@
+import { setupDictionaries } from './core';
+import { dictionariesSerializer } from './serializer';
+
+export const dictionaries = {
+	serializer: dictionariesSerializer,
+	functions: {
+		setupDictionaries,
+	},
+};
+
+export * from './types';
