+++ packages/codegen/src/core/modules/JavaScript/JavaScriptCompiler/dictionaries/types.ts
@@ -1,0 +1,1 @@
+export type TDictionaries = string[];
