+++ packages/codegen/src/core/modules/JavaScript/JavaScriptCompiler/expressions/index.ts
@@ -1,0 +1,12 @@
+import { getExpressionValue, setupExpressions } from './core';
+import { registerCustomFunctions } from './functions';
+import { expressionsSerializer } from './serializer';
+
+export const expressions = {
+	serializer: expressionsSerializer,
+	functions: {
+		getExpressionValue,
+		registerCustomFunctions,
+		setupExpressions,
+	},
+};
