+++ packages/codegen/src/core/modules/JavaScript/JavaScriptCompiler/imports/index.ts
@@ -1,0 +1,12 @@
+import { buildDependencyGraph, checkForCyclicDependencies } from './core';
+import { importsSerializer } from './serializer';
+
+export * from './types';
+
+export const imports = {
+	serializer: importsSerializer,
+	functions: {
+		checkForCyclicDependencies,
+		buildDependencyGraph,
+	},
+};
