+++ packages/codegen/src/core/modules/JavaScript/JavaScriptCompiler/imports/types.ts
@@ -1,0 +1,4 @@
+export type TImports = {
+	[path: string]: string[];
+};
+export type TDependencyGraph = Map<string, Set<string>>;
