+++ packages/codegen/src/core/modules/JavaScript/JavaScriptCompiler/index.ts
@@ -1,0 +1,15 @@
+import { classModule } from './class';
+import { context } from './context';
+import { dictionaries } from './dictionaries';
+import { expressions } from './expressions';
+import { imports } from './imports';
+import { state } from './state';
+
+export const JavaScriptCompiler = {
+	class: classModule,
+	context,
+	dictionaries,
+	expressions,
+	imports,
+	state,
+};
