+++ packages/codegen/src/core/modules/JavaScript/JavaScriptCompiler/state/index.ts
@@ -1,0 +1,7 @@
+import { getInitialState } from './core';
+
+export const state = {
+	functions: {
+		getInitialState,
+	},
+};
