--- packages/codegen/src/core/modules/JavaScript/imports.ts
@@ -1,13 +1,0 @@
-// export type TImports = {
-// 	[path: string]: string[];
-// };
-//
-// export function getImports(props: {
-// 	imports: TImports;
-// }) {
-// 	let imports = '';
-// 	for (const [key, value] of Object.entries(props.imports)) {
-// 		imports += `import { ${value.join(', ')} } from '${key}';\n`;
-// 	}
-// 	return imports;
-// }
