+++ packages/codegen/src/core/modules/JavaScript/index.ts
@@ -1,0 +1,7 @@
+import { JavaScriptCodegen } from './codegen';
+import { JavaScriptCompiler } from './JavaScriptCompiler';
+
+export {
+	JavaScriptCodegen,
+	JavaScriptCompiler,
+};
