--- packages/codegen/src/core/modules/JavaScript/shorties.ts
@@ -1,7 +1,0 @@
-// export function getReferenceString(path: string, identifier: string) {
-// 	return `${path}['${identifier}']`;
-// }
-//
-// export function getFunctionFromDictionary(name: string) {
-// 	return `functionDictionary.get('${name}')`;
-// }
