--- packages/codegen/src/core/modules/JavaScript/types.ts
@@ -1,5 +1,0 @@
-// import { TConstants } from '../../../types/common';
-//
-// export type TJavaScriptThis = {
-// 	constants: TConstants | null;
-// };
