+++ packages/codegen/src/core/modules/TypeScript/TypescriptCompiler/class/index.ts
@@ -1,0 +1,5 @@
+import { JavaScriptCompiler } from '../../../JavaScript';
+
+export const classModule = {
+	serializer: JavaScriptCompiler.class.serializer,
+};
