+++ packages/codegen/src/core/modules/TypeScript/TypescriptCompiler/index.ts
@@ -1,0 +1,11 @@
+import { JavaScriptCompiler } from '../../JavaScript';
+import { classModule } from './class';
+
+export const TypeScriptCompiler = {
+	class: classModule,
+	context: JavaScriptCompiler.context,
+	dictionaries: JavaScriptCompiler.dictionaries,
+	expressions: JavaScriptCompiler.expressions,
+	imports: JavaScriptCompiler.imports,
+	state: JavaScriptCompiler.state,
+};
