+++ packages/codegen/src/core/modules/TypeScript/index.ts
@@ -1,0 +1,5 @@
+import { TypeScriptCodegen } from './codegen';
+
+export {
+	TypeScriptCodegen,
+};
